+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/CheckMojo.java
@@ -1,6 +1,5 @@
 package nbbrd.heylogs.maven.plugin;
 
-import internal.heylogs.StylishFormat;
 import internal.heylogs.semver.SemVer;
 import nbbrd.heylogs.Check;
 import nbbrd.heylogs.Heylogs;
@@ -11,27 +10,27 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 
 import java.io.File;
+import java.io.IOException;
+import java.io.Writer;
 
+import static internal.heylogs.maven.plugin.HeylogsParameters.*;
 import static java.util.Collections.singletonList;
 import static java.util.Locale.ROOT;
 
 @Mojo(name = "check", defaultPhase = LifecyclePhase.VALIDATE, threadSafe = true)
 public final class CheckMojo extends HeylogsMojo {
 
-    @Parameter(defaultValue = "${project.basedir}/CHANGELOG.md", property = "heylogs.input.file")
+    @Parameter(defaultValue = DEFAULT_INPUT_FILE, property = INPUT_FILE_PROPERTY)
     private File inputFile;
+
+    @Parameter(defaultValue = DEFAULT_OUTPUT_FILE, property = OUTPUT_FILE_PROPERTY)
+    private File outputFile;
 
     @Parameter(defaultValue = "false", property = "heylogs.semver")
     private boolean semver;
 
-    @Parameter(defaultValue = StylishFormat.ID, property = "heylogs.format.id")
+    @Parameter(defaultValue = DEFAULT_FORMAT_ID, property = FORMAT_ID_PROPERTY)
     private String formatId;
-
-    @Parameter(defaultValue = "${project.version}", readonly = true)
-    private String projectVersion;
-
-    @Parameter(defaultValue = "${project.basedir}", readonly = true)
-    private File projectBaseDir;
 
     @Override
     public void execute() throws MojoExecutionException {
@@ -67,12 +66,19 @@ public final class CheckMojo extends HeylogsMojo {
 
     private void check() throws MojoExecutionException {
         Heylogs heylogs = initHeylogs(semver);
+
         Check check = Check
                 .builder()
                 .source(inputFile.toString())
                 .problems(heylogs.validate(readChangelog(inputFile)))
                 .build();
-        log(text -> heylogs.formatProblems(formatId, text, singletonList(check)), check.hasErrors() ? getLog()::error : getLog()::info);
+
+        try (Writer writer = newWriter(outputFile, check.hasErrors() ? getLog()::error : getLog()::info)) {
+            heylogs.formatProblems(formatId, writer, singletonList(check));
+        } catch (IOException ex) {
+            throw new MojoExecutionException("Error while writing", ex);
+        }
+
         if (check.hasErrors()) {
             throw new MojoExecutionException("Invalid changelog");
         }