+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/ListMojo.java
@@ -1,19 +1,27 @@
 package nbbrd.heylogs.maven.plugin;
 
-import internal.heylogs.StylishFormat;
 import nbbrd.heylogs.Heylogs;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 
+import java.io.File;
+import java.io.IOException;
+import java.io.Writer;
+
+import static internal.heylogs.maven.plugin.HeylogsParameters.*;
+
 @Mojo(name = "list", defaultPhase = LifecyclePhase.VALIDATE, threadSafe = true)
 public final class ListMojo extends HeylogsMojo {
+
+    @Parameter(defaultValue = DEFAULT_OUTPUT_FILE, property = OUTPUT_FILE_PROPERTY)
+    private File outputFile;
 
     @Parameter(defaultValue = "false", property = "heylogs.semver")
     private boolean semver;
 
-    @Parameter(defaultValue = StylishFormat.ID, property = "heylogs.format.id")
+    @Parameter(defaultValue = DEFAULT_FORMAT_ID, property = FORMAT_ID_PROPERTY)
     private String formatId;
 
     @Override
@@ -28,6 +36,11 @@ public final class ListMojo extends HeylogsMojo {
 
     private void list() throws MojoExecutionException {
         Heylogs heylogs = initHeylogs(semver);
-        log(appendable -> heylogs.formatResources(formatId, appendable, heylogs.getResources()), getLog()::info);
+
+        try (Writer writer = newWriter(outputFile, getLog()::info)) {
+            heylogs.formatResources(formatId, writer, heylogs.getResources());
+        } catch (IOException ex) {
+            throw new MojoExecutionException("Error while writing", ex);
+        }
     }
 }